feat: create a folder (#52)
