+++ src/controllers/authentication/login/_handler.py
@@ -1,13 +1,17 @@
+import json
 from flask import request
 from src.config.soap_client import soap_client
 
 
 def login_handler():
     try:
-        data = request.json
+        data = json.loads(request.data)
         username = data.get("username")
         password = data.get("password")
-        if not username or not password:
+
+        not_valid_username = not username or len(username) == 0
+        not_valid_password = not password or len(password) == 0
+        if not_valid_username or not_valid_password:
             return {"msg": "Required fields are missing in JSON data"}, 400
 
         result = soap_client.service.auth_login(
@@ -20,23 +24,9 @@ def login_handler():
 
         return {"msg": "Invalid credentials"}, 401
 
+    except ValueError:
+        return {"msg": "Invalid JSON data provided in the request"}, 400
+
     except Exception as e:
         print("[Exception] login_handler ->", str(e))
         return {"msg": "There was an error logging in"}, 500
-
-
-def challenge_handler(token):
-    try:
-        response = soap_client.service.auth_refresh({"token": token})
-        if hasattr(response.auth, "token"):
-            refreshed_token = response.auth.token
-            return {
-                "msg": "JWT refreshed successfully",
-                "token": refreshed_token,
-            }, 200
-        else:
-            return {"msg": response.msg}, response.code
-
-    except Exception as e:
-        print("[Exception] challenge ->", str(e))
-        return {"msg": "There was an error validating or refreshing the session"}, 500
